tasks/cephfs: refactor test case class
A more generic CephTestCase and CephCluster, for
writeing non-cephfs test cases.

This avoids overloading one class with the functionality
needed by lots of different subsystems.
